+++ Assets/Demand/KnownPrice.cs
@@ -1,5 +1,5 @@
 public class KnownPrice
 {
     public Good Good;
-    public decimal Price;
+    public int Price;
 }
