+++ Assets/FixedCosts/iFixedCostStrategy.cs
@@ -2,6 +2,6 @@ using System.Collections.Generic;
 
 public interface iFixedCostStrategy
 {
-    decimal CalculateFixedCosts(List<FixedCostInstance> fixedCosts, int period);
+    int CalculateFixedCosts(List<FixedCostInstance> fixedCosts, int period);
     void SetEconAgent(EconAgent agent);
 }
