+++ Assets/Markets/AvailableGood.cs
@@ -3,9 +3,9 @@ public struct AvailableGood
     public iEconAgent Agent;
     public Good Good;
     public int Quantity;
-    public decimal Price;
+    public int Price;
 
-    public AvailableGood (iEconAgent agent, Good good, int quantity, decimal price)
+    public AvailableGood (iEconAgent agent, Good good, int quantity, int price)
     {
         Agent = agent;
         Good = good;
